03_SQL_Pytest/requirements.txt
pytest==8.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1